// src/app/shell/enums/chart.types.ts
export enum ChartTypes {
  VerticalBar = 1,
  HorizontalBar = 2,
  Ring = 3,
  Line = 4,
  KPI = 5,
  StackedHorizontalBar = 6,
  StackedVerticalBar = 7,
  MeanChart = 8,
  BarChart = 9,
  Table = 10,
  Scatter = 11,
  Pie = 12,
  Area = 13,
  Bubble = 14,
  Text = 15,
  Gauge = 16,
  Radar = 17,
  Profile = 18
}